microsites/query_service: enable TLS when monitoring commons-query
Back in March 2023 when we added http blackbox monitoring for
all sites / virtual hosts we run on miscweb* and elsewhere,
I had noticed that TLS didn't work for:

https://commons-query.wikimedia.org

as opposed to other services such as https://query.wikidata.org.

I said back then the host name had not been added to the
envoy TLS cert yet and made it use plain http and changed port to 80.
(Ia91eaf916f4a1731 et al)

Now I can confirm it is an alt_name on the cert and it is
part of this cleanup task to fix this.

So optimistically enabling TLS for monitoring to confirm it
actually works now.

Compare to monitoring for query.wikidata.org in this same file.

Bug: T333510
Change-Id: I9bf673fe7ae28860e81a532208c05e23624a35bc