varnish: enrich X-Analytics for browser prefetch / prerender / preview
This tags for known patterns for headers:

 - Chrome / Chromium / modern Edge (Sec-Purpose: and Purpose:)
 - Safari / WebKit (Purpose:)
 - Firefox pre v. 115 (X-Moz:) and 115+ (Sec-Purpose:)

It also notes when an unknown Sec-Purpose: header value is encountered.

It's not possible to know what formats the Sec-Purpose: header will
take in the future, but for now specific header formats are mostly
anticipated. The header format has semicolons in practice in Chrome /
Chromium, and instead of transcoding the semicolon (which is an actual
delimiter in the X-Analytics header) or creating yet another header
we opt here to handle the common cases and defer on treating other
unknown cases until the need arises.

For Chrome

Bug: T346463
Change-Id: I7c8b33eca0891725f28250c9f0c54db7206182a5