admin: allow using security key backed SSH keys
On OpenSSH 8.4 and newer (introduced in Debian bullseye) there is native
support for FIDO security key backed keys, which is much more convinient
than the current GPG workaround. This patch adds support for specifying
certain keys on Buster only, so that I can add such a key to my account
and eventually only have that for hosts where it's supported and still
use the "old" key for logging in to buster hosts.

Change-Id: Icd5156db9159ad816556b219aca930da3c21e96b