defs_requestctl_nftables.tpl: simplify template
Why:
Given we're now only querying for a single key, the template can be
simplified.

What:
Use getv to get directly the value and hardcode the define name that is
then used by nftables so that even a rename in the etcd data would not
break it.
Get rid of the unnecessary newlines and spaces too.

Hosts: sretest1001.eqiad.wmnet, mw1399.eqiad.wmnet
Hosts: idp1002.wikimedia.org, bast4005.wikimedia.org
Bug: T348734
Change-Id: I1bb8a5b5955286e52e531b8aa8a1e6eb6da83b94