memcached: provide the CA cert when listening to TLS
When listening to TLS, memcached is using the server's puppet
certificate. Puppet 7 hosts use a new CA, however, this CA cert
is not present on puppet 5 hosts, yielding a SSL certificate error.

This patch allows us to put a bandaid by presenting the CA cert
to the clients, until we fully migrate to PKI.

This is a noop, we will define profile::memcached::localcacert in
future patches.

Bug: T349619
Change-Id: I5e20548d9bf3cefe5c7f41b3a2ae87243bbd8636