C:puppetmaster::monitoring Prometheus stats for Puppetmerge.
For us to move Puppetmerge monitoring to Prometheus, we need to start
collecting the required data, missing merges.

This adds a new script for handling the exporters. It is based on the
existing shell script, but removed some of the redundancy in the old
version. The script write a prom-file in the node.d directory, allowing
the data points to be collected by the node exporter, which is already
running.

Comment from the old Icinga implementation claims to run every minute,
but looking at the module it seems to in fact be running every 10
minutes.

Bug: T350694
Change-Id: Id6870cb2eb96540ffe5e272fe3c193388420a791