yarn: configure Apache to only listen to port 80
We need to enable mod_ssl to proxy requests over to a service
via https. However, in the current state of configuration, enabling
mod_ssl causes apache2 to listen to the port 443 as well, which is
already listened to by envoy.

By setting `http_only` to `true`, we instruct apache to only listen
to port 80, even if mod_ssl is loaded (cf T277989).

Bug: T352863
Hosts: an-tool1008.eqiad.wmnet
Change-Id: I56920e316cf14efc9bace8c610f0089b7e812e24