wdqs ldf: remove useragent param
We are the only blackbox check using the useragent
param. (Confirmed by looking in puppet repo and
/etc/prometheus/blackbox.yml.d/ on prom host).

Remove this param to make it possible to
completely remove it from the Puppet module.

This is a first step down the path of exposing
the ability to set arbitrary headers in our
prometheus::blackbox::check::http Puppet resource.

Hosts: prometheus1006.eqiad.wmnet
Bug: T353672
Change-Id: Ie8c3e4854b2b1ab7226e2a0598334df67d54dc15