phabricator: move enable_vcs parameter from eqiad-only to role
git-ssh.wikimedia.org has been deprecated and since then
we don't configure an extra IP address for sshd anymore.

This was the reason to do this only in one DC, to avoid
setting the extra IP on 2 hosts.

We don't do this anymore but need to keep the rest
of vcs enabled.

So now we can move this to role level to make it less
confusing and avoid a potential surprise if we ever
switch to codfw and forget about this.

Bug: T296022
Change-Id: Ie675904b8004e8e6b3ae2d12874a37f7c495ce36