services_proxy: Switch listen_ipv6 to true by default
Why:
In T255568, back in 2021 we introduced the ability for users of the
services_proxy to listen on IPv6 and IPv4. We set those for various
MediaWiki clusters. It's about time to use it as a default

What:
Delete all hiera keys set to true and switch the default from false to
true. Add a comment to explain the exclusion of restbase

Hosts: auto
Bug: T255568
Change-Id: I1c5cd52c4cc797696199d33b0a1b6083bea41464