httpd-yarn: proxy reqs with a /spark-history prefix to the spark-history svc
The /history prefix is the one used by yarn to indicate that we're
hitting a spark history page, so this should hit
<spark-hist-svc>/history. However, we also need to serve the spark
hsitory statics.

In
https://gerrit.wikimedia.org/r/c/operations/deployment-charts/+/984127,
we configured spark to prepend all its links with /spark-history, to
inform the apache reverse proxy that requests should be proxied to it,
which is what this patch does.

Bug: T352863
Change-Id: I9ef5f03f248e7e6bff44272c1de695fe4dbec1b3