Retrict access to the spark-history k8s API tokens
We are supplying a kerberos keytab to the spark-history deployments,
which could be used to access HDFS as a member of the
analytics-privatedata-users group.

Given that the default ownership of the files containing the API tokens
for these namespaces is mwdeploy:deployment, there is the possibility
that a member of the deployment group could elevate their privileges to
a member of the analytics-privatedata-users group by retriving the
keytab.

For this reason, we wish to retrict access to these API tokens to
members of the analytics-deployers group, who already have the level of
access.

Hosts: O:deployment_server::kubernetes
Bug: T330176
Change-Id: Ia74de476df85104b150d7272a2f018b2306a1ffb