envoy-build-config: call extend() instead of append() if passed a list
OK, this is going weird but this change isn't really needed. append()
will anyway DTRT and create a nested list. Which is BOTH valid yaml AND
valid envoy configuration. In fact, envoy will apparently DTRT and honor
it properly in my limited tests.

However, having weird stanzas like the one listed in

https://phabricator.wikimedia.org/P54492

in our configuration is bound to lead to confusion some poor soul, the
author of this patch included.

What:
Detect if we read a list and then call extend() instead of append()

Bug: T255568
Change-Id: I6cf43efb701ff687fdce5995ba17713c0c8fbfd4