Netbox report, reduce alerting spam
A bug somewhere causes the Netbox report to occasionnaly error out this
causes a significant amount of IRC spam on the -dcops channel.
This worksaround the issue by only alerting on "failed" reports and thus
not "errored" ones.

To be rolled back after the root cause of the error is figured out.
Or at least a test after the next Netbox upgrade.

Bug: T321704
Change-Id: Ic32e9ccd9c2323691b06b0eee069ac3fc721e952