wdqs: graph split hosts don't need categories
We debated between forking off a new
graph_split_test role versus
changing the existing test role.

We're going to just use the test role. If we need
categories on future (non-graph-split) test hosts
we can just revert this patch.

---

We're not positive, but we think categories systemd
failures might have contributed to the issue
in the below bug.

Hosts: wdqs1022.eqiad.wmnet
Bug: T352878
Change-Id: I4391606043b8f9319e6b1f0b7b8b84e0d8e5a475