RESTBase: Hardcode no_worker patch to stop OOM
Why:
We 've been witnessing Cassandra getting OOMKilled on newer sets of
hosts that have more CPUs than the older set of hosts (64 vs 40). The
extra set of CPUs means restbase spawns more workers, thereby increasing
the overall memory usage, forcing OOMKiller to show up. OOMKiller
decides that the best candidate is one of the Cassandra instances and
kills it

What:
Add a hiera lookup to allow configuring the number of workers. Supply it
with a default of ncpu, then override it just for production with a
value of 40

Hosts: auto
Bug: T353456
Change-Id: Ida4c983169ea5f218357733372217f440f1979ec