Bring dbstore1008 into service
dbstore1008 is the new bare-metal server to replace dbstore1003 as an
analytics replica mariadb server. We wish to begin the process of
migrating the service from the older to the newer host.

In addition, we switch the remaining dbstore1009 host as being
the responsibility of the data engineering team, rather than the data
persistence team.

Bug: T351921
Change-Id: Ief37099c5707def338ce06461dc538e7313f9efc