global_config: list IPs of hadoop master/workers and kerberos nodes
We include the IPs of hadoop masters, hadoop workers and kerberos nodes
to the side global config files rendered on the deployement server.

Coupled with https://gerrit.wikimedia.org/r/c/operations/deployment-charts/+/984819
this will allow us to define Kubernetes Services/Endpoints for these nodes
that will be then able to be referenced in Calico NetworkPolicies,
to allow egress to these nodes.

This will improve the current situation where we allow egress to some nodes
on all the DPE subnets, as a way to circumvent the fact that the precise IPs
were not available to helmfile.

Bug: T331894
Hosts: deploy2002.codfw.wmnet
Change-Id: I1918556b40ba3e2a766fad5beec99f57db3aba76