raid handler: fix broken cases
Why:
The current code for the raid handler is half broken depending on the
combination of output and if the raid output is compressed or not.
The current situation was caused by the combination of those two
patches:
  * I7d210ba4adcc728e7836124db2f2fd09d7bdb19d
  * I709e8f177f8e712210497b11f5b9b36f7ecda235

Neither is the correct solution because from one side we need the output
to be binary for the compressed cases (so universal_newlines/text must
be False) and from the other side decode() leaves newlines encoded. We
could use decode('unicode_escape') but that would cause other encoding
issues.

Example of errors in the logs:

1) Failed to decompress:
[ERROR] raid_handler::get_raid_status: Failed to execute '['/usr/lib/nagios/plugins/check_nrpe',
  '-4', '-H', 'db1168', '-c', 'get_raid_status_megacli']': 'utf-8' codec can't decode byte 0x9c in
  position 1: invalid start byte

2) Failed to decode a string
2024-01-03 05:24:51 [ERROR] raid_handler::<module>: Unable to handle RAID check alert
Traceback (most recent call last):
  File "/usr/lib/nagios/plugins/eventhandlers/raid_handler", line 282, in <module>
    main()
  File "/usr/lib/nagios/plugins/eventhandlers/raid_handler", line 259, in main
    status=get_raid_status(args.host_address, args.raid_type))
  File "/usr/lib/nagios/plugins/eventhandlers/raid_handler", line 122, in get_raid_status
    status = stdout.decode()
AttributeError: 'str' object has no attribute 'decode'

What:
Use text=True when dealing with an uncompressed raid type and False
otherwise. Remove the call to decode() in case of uncompressed raid type
as it will be already a string.
Tested generating missing tasks for the latest failures:
* T354250 (perccli)
* T354251 (md)
* T354249 (md)
* T354253 (megacli) # disk already fixed hence the optimal status

Change-Id: Iee46771baac911e4a3005e91ab9ef68a55d06a99
Hosts: auto