mwlog: Switch rsync service to use firewall::service
1930b5c213e7703e53862b345f5dbe9b415bb6c0 added a new option to the
rsync::quickdatacopy define to add a firewall::service definition
for the included firewall rule instead of a ferm-specific definition.

This will eventually allow to migrate invididual roles to nftables.

This slightly changes the behaviour, but in a good, intended way:
The only variant explicitly disabled Ferm AAAA lookups via auto_ferm_ipv6,
but with the new approach used in firewall::server the host names are
resolved on the Puppet server (and thus return ipv4 and/op ipv6 depending
on what is stored in DNS).

Hosts: mwlog1002.eqiad.wmnet,mwlog2002.codfw.wmnet
Change-Id: Ib81a9aa451b7af006d095d62dd8513de70d995d2