swift: add new nodes, drain old nodes from the rings
We are doing a hardware refresh of swift backends. That means we're
draining 14 old nodes:
ms-be10[44-50]
ms-be20[44-50]

and adding 14 new nodes:
ms-be10[76-82]
ms-be20[74-80]

It's a historical artifact that we have uneven numbers of servers in
the two ms clusters...

This change will then be automatically rolled out gradually by the
ring management code over a period of a few weeks.

Bug: T353149
Change-Id: If0933dc8f4a5b555fa620a6262db030237f57a3e