services_proxy: Support tracing
Why:
We want to enable tracing in the services proxy as well.

For now we want to see if we can enable it for all parts of the services
proxy in a somewhat easy way, via a simple hiera key. Further down the
line, we probably can switch that to per listener, but let's prove this
work first

What:
* Check is we have a hardcoded upstream otel-reporter envoy cluster. If
  not, define one. Do the same thing in the tlsproxy too so that we
  end up with duplicate definitions
* Amend the services_proxy listener to enable, based on a hiera variable
  random sampling for all listeners
* Populate the required hiera keys for all cases

Hosts: mwdebug1001.eqiad.wmnet
Hosts: mw1356.eqiad.wmnet
Bug: T351566
Change-Id: I36226368d51a7f4ef4015af35f0011d6d072553f