phabricator: use quickdatacopy for automatic home dir sync
After I796d928b2d5 et all we now have a single setting each
for the active and the passive phab/phorge server so we
can use that as src and dest hosts.

Then use quickdatacopy to automate the syncing of home dirs
between the 2 where the active server is always the source.

Bug: T354221
Change-Id: I4924da46c3a6dd9493506881b27162a75aa840af