[gerrit] Refactor classes to specify an active host
This replaces the `src_host` and `dst_hosts` from `migration.pp` and the
`is_replica` flags from a number of other classes. We replace this with
an `active_host` flag in the `gerrit.yaml` config, and then build a list
of destination hosts from the `wmflib::class::hosts` call to find all
hosts that have the gerrit flag applied.

Change-Id: I73c9b6395e9c859495531c4df9c6c760ebc0df29