dologmsg: standardize logging format
Include user@hostname instead of <user>, to uniform this to the logging
format used in other places and parsed by stashbot.

Also make sure this cannot be run from anywhere that is not a Toolforge
bastion, as it would result in an invalid log message.

Bug: T346631
Hosts: tools-sgebastion-10.tools.eqiad1.wikimedia.cloud
Change-Id: I71aed7345702cfe9be7ecf5e78d879ba0212c9cc