phabricator: avoid duplicate lists of servers in Hiera
Instead of listing the phabricator servers in multiple
places define the passive server in common Hiera alongside the active
server and use this across the board.

Where the list of all servers is needed just replace
it with a string combined from both servers.

This way we can get rid of $phabricator_servers as well as
$phabricator_servers_ferm and have made things a little
bit simpler.

Also rename "phabricator_server" to "phabricator_active_server"
and failover server to "phabricator_passive_server" to have only
one set of names across the board.

Bug: T354221
Change-Id: I796d928b2d57ccd431ac5041469351feb7d2d20d