phabricator: use same db server regardless of DC of phab server
In the past we expected that the eqiad phab/phorge server should talk
to the eqiad database server and the codwf phab/phorge server should talk
to the codfw database server.

But after recently discussing failover options and plans again we now
thin we want to always talk to the same database regardless of the DC
of the phab/phorge server.

This is because in the event of a failure of the phab/phorge server
we could more easily failover to the other server without having to
contact DBA and change which database server is writable and which is
not.

Also, while still not ideal, cross-DC database queries are now
considered acceptable for a limited time.

In that limited time we could repair or replace the original server.

We still need to be sure that we never write to the DB from 2 servers
at a time though and somehow ensure that!

Bug: T334519
Change-Id: I435c394f4c565808d8d33fd3b9cd7e5a0ba25cc6