phabricator: temp test of repo syncing, using gitlab2003 spare host
This is for a one-off test to see how long syncing actually takes.

gitlab2003 is being used because it is currently not in production
and has enough disk space.

Will be reverted right after the test.

But doing this in puppet to keep everything clean without manual
changes to the server.

Bug: T334519
Change-Id: I25125303f12743e8ce6784aece9bba20eeb0fa6e