prometheus::blackbox::check: make for parameter configurable
As discussed in T354479 some services with lower SLA should alert only
if the service is down for an extended period of time.
To implement this Prometheus 'for' clause can be used.
This change adds a optional parameter to the blackbox::check::http and
blackbox:check::tcp. Default is 2m so it should not affect existing
blackbox checks. But additional attention to diffs of existing
checks might be a good idea.

Bug: T354479
Change-Id: I1185020dbc32b3efa16e5cc83bd6542505f31bda