replica_cnf_api: Do not check for file existence
If a file exists, either it is up-to-date (in which case we want to tell
maintain-dbusers everything is ok so it stops retrying), or it is
outdated, in which case it needs to be overwritten to have the correct
values.

Additionally the old code would crash if the tool home directory was not
readable.

Bug: T355356
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I0e06e6c986dd043becb552d5a1b6b7642ae4cf1c