Fix deployment diff issue and clean up presentation
The diff portion of the deployment script was not behaving as desired.
The grep -v "./" command was causing any file in a subdir to match and
thus hiding changes when trying to verify what will be done. This change
escapes the . to make it a literal match.

Also updated the deploy step to run puppet after the file copy. This is
to fix up any possible permission issues on the files/dirs and any
templated files in the off chance they get updated.

Additionally, updated the help and some other echo statements to make
the output a little cleaner and consistent.

Bug: T343486
Change-Id: I7cd31dfe20b3d2708ace6bf0c05b7ecbf7992cc8