archiva: fix 400s when proxying requests with spaces in the URL
We've seen that calls to
https://archiva.wikimedia.org/restServices/.../Repository%20Observer%20-%20python
fails with an empty 400 response, but a call to
https://archiva.wikimedia.org/restServices/.../Guests
succeeds.

After a bit of tinkering, I found that this is due to nginx preventing
the proxy passing of requests with spaces in the URL
(cf https://trac.nginx.org/nginx/ticket/2373).

The suggested solution is to omit the URI component in the proxy pass directive,
causing nginx to pass URIs from the client request unmodified (and properly escaped).

A manual edit + reload showed that this configuration worked.
I had to also remove the trailing slash, as keeping it caused a syntax validation
of the configuration file.

Bug: T355352
Hosts: archiva1002.wikimedia.org
Change-Id: Iad5a451b459e52cb25926f766dfdf22af756691b