nova policy: add awareness of 'unmanaged' role
For now this role will allow a user to 'rebuild' VMs. Rebuilding
doesn't work with puppetized VMs (because the cert doesn't get reset)
but may be useful to users who have already foregone puppet.

Right now I need this because I want the Trove service user to be
able to rebuild VMs as part of the database instance rebuild
process.

Bug: T326818
Bug: T355721
Change-Id: I910ebdf903b31e943575ce9b498e99b07a849796