gerrit: use finer groups for commit commentlink
d618f0c8da8 made the commit commentlink to use a `link` but missed that
the text of the HTML anchor defaults to the whole matched text ($&).

Since the regular expression looks for a leading sequence before the
commit sha1, it shows up in the HTML anchor:

  <a href="/q/abcdef"> abcdef</a>
                     ^^^

Explicitly set the link text to the commit sha ($2) instead of using the
whole match ($&).
Add the leading sequence ($1) in front of the anchor with `prefix`.

Bug: T354886
Bug: T355537
Change-Id: If61a97a01f9b68ff36084836cdb355787baf7e3a