sql/xml dumps: add role for helper worker for wikidata full history dumps
This will run a separate backfill script to cover parts near the end of
the full history job, on a separate worker host.
The run will take place
after the main dump worker should have produced the input files
necessary (stub files with small page ranges, and the pagerange.json
file with history information). It should complete well before the main
dump worker would start to process these parts near the end of the job
itself.
We want to avoid overlap because there is no per-job locking,
let alone per-file, and implementing that would be too costly for a
system that is being replaced. Nonetheless, the backfill script takes
some measures to avoid writing over output files produced by the main
worker.
Scheduling was decided manually after checking when various input and
output files are written by the main worker. The schedule is set in the
hiera file for the snapshot host that will run the backfill script
(range of parts to be done goes here) and in the dump_fillin_wd class
manifest (dates this script should run or retry).

Hosts: snapshot1015.eqiad.wmnet
Bug: T252396
Change-Id: Ia96698aa70f5e5f710725757eaacfffc6ae6eca7