airflow: Keep Python 2 due to Hive
All Airflow roles include Hive in profile::analytics::cluster::airflow.
Hive depends on Python 2, which we deinstall by default on Bullseye.
As such, Puppet on Airflow Bullseye hosts is in a catch-22 between the
base class deinstalling Python and then Hive pulling it back in:

 Notice: /Stage[main]/Base::Standard_packages/
    Package[libpython2.7-minimal]/ensure: removed (corrective)
 Notice: /Stage[main]/Base::Standard_packages/
    Package[python2.7]/ensure: removed (corrective)
 Notice: /Stage[main]/Base::Standard_packages/
    Package[libpython2.7-stdlib]/ensure: removed (corrective)
 Notice: /Stage[main]/Base::Standard_packages/
    Package[python2.7-minimal]/ensure: removed (corrective)

As such, set the override flag for these roles to keep Python 2 installed.

Bug: T335261
Change-Id: Iceb814815fe78ec25eaf6dc9f84c29a1121803df