reposync: don't enforce ownership after init
Prior to this change 'recurse' was set on the repo directory. When set
puppet creates file resources for every file in the direcgtory. On
netbox1002 this caused warnings about exceeding the soft limit of 1000
and causing performance degradation. The recurse parameter performs
poorly with a large number of files.

After this change we remove the recurse parameter and add the group
parameter on the git init call. The tested result is that new repos have
the correct user and group applied. Now however, changing the group or
user after the fact has no effect on the repos content. In addition
puppet will no longer fix up the user and group if the git pulls are
performed under the wrong user or group. This trade off seems
acceptable.

Hosts: C:reposync
Change-Id: Id4ed175f60991d410d3bf5630eb0d249e60372bb