fifo-log-demux: Decouple service from nginx/ats
Deploying a change to the systemd unit for fifo-log-demux caused nginx
to be restarted on the ncredir servers. fifo-log-demux should be able to
be restarted without bringing nginx down with it.

However, it's important to ensure to keep fifo-log-demux running as
some services will stall out while waiting for a process to consume the
named pipe. To that end, set up the service to attempt restarts
eternally at a continual rate in an attempt to salvage operations
without suffering the default time increases in service restarts.

Bug: T355905
Change-Id: Ifc6ab1f2bf507ddbaf60ff5c70ee8605bae107a7