DHCP: set "use-host-decl-names on"
so I forgot to add the PTR include to the DNS zone for a new IP prefix
and DHCP sends option 12 (hostname) in its reply along with the IP
but to figure out what the hostname is, DHCP currently sends a DNS query
on the IP.
and D-I uses that hostname to decide which preseed config to apply...

So the hostname was set to "undefined" and DI wasn't working.

The obvious fix is to add the missing PTR include to DNS.

But it's also silly to depend on DNS while the data is already present
in the DHCP config.

Manually tested and works fine.

Bug: T300152
Change-Id: Ib9c94b0a327fd942ffb322fac190694031156da5