P:debmonitor::server Do not embed the fqdn of an instance in rules.
We cannot use FQDNs as server_name for blackbox monitoring, as this will
embed the FQDN in the blackbox rules on the Prometheus hosts. This will
query most instances with the wrong host header.

This is a result of the blackbox exporter running on the prometheus
hosts, as they should, and not on the individual hosts as I wrongly
assumed.

Change-Id: Id69b81a310c578c4675b408f3776c1fbc5a65b63