netboxdb: change bacula settings for DB backup
Why:
The current setting is backing up the whole directory where we do create
a new file each day and then rotate them, and that's inefficient and
resource wasting for bacula.

What:
Now that a fixed file name is always hard linked to the latest backup we
can tell bacula to just backup that single file, and doing it daily,
to be more efficient.

Hosts: auto
Bug: T316655
Change-Id: Iad93eb3cf3e491557189319359170bd4138094a8