firewall: fix nftables metric exporter
Why:
The current script to export prometheus metric for the existence of the
firewall outputs the file for prometheus-node-exporter with the wrong
format, causing lots of errors to be logged on the hosts.

What:
Fix the script with the proper quotes around the metric label and also
interrupt the execution in case of error without trying to parse a
variable that is not set.

Hosts: auto
Bug: T356305
Bug: T348499
Change-Id: I0424341702186a9b080d7ee7c276488f155830c7