requestctl-generator: fix bug for URI filters
Why:
Requestctl filters on URI work on the whole URI while in the superset
dashboard the URI path and query filters are separate.
The requestctl-generator has a bug when generating rules for all
combinations of filtered URI paths and URI queries that if one of the
two set is empty (e.g. the user has filtered only by path and not
query), would not generate the correct rule for requestctl.

What:
Set the URI query set as made by the emoty query if there are URI path
filters but no URI query and viceversa to ensure the proper filter are
always generated.

Hosts: auto
Change-Id: Id6f61b750b76a845ed39b938b187f7600b70432c