cassandra: break-fix image suggestions Cassandra access
The image suggestions DAG (airflow) made use of the aqsloader role
instead of the one that was created for it (`image_suggestions`).
When c47109a was applied (see: T355917), the ALL KEYSPACES GRANTs
were removed and replaced with the specific/applicable list of
keyspaces.  This broke the image suggestions jobs.

This restores the GRANTs to `aqsloader` as a temporary break-fix.

Bug: T356400
Change-Id: Ia89259faf6b5960e5b182cc04a42ee4d2d1fcfe6