P:httpbb: migrate tests from cumin1001 to cumin1002
To support this, the state of the ensure property on the k8s test timer
resources is now unconditionally controlled by test_kubernetes_hourly.

Notably, this means that there will be long-term absented resources on
hosts with profile::httpbb.

The way in which ensure is controlled for the non-k8s test timers does
not change, and still involves a two-phase cleanup (i.e., set host list
to empty then remove).

Hosts: auto
Bug: T356054
Change-Id: I8db6311d42dc0cc4fcb1fc7a044bda1185e9f1cb