Temporarily enable Dockerfile frontend on trusted runners
Temporarily allow the Dockerfile (dockerfile.v0) frontend on trusted
runners to allow Release Engineering to publish a new buildkitd
container image for a security release.

We'll need to find a better way to deal with this situation in the
future.  Converting buildkitd's Dockerfile to blubber format is not
feasible, and there are others who would benefit from allowing the
Dockerfile frontend on trusted runners (T351792).

Bug: T356418
Change-Id: Iab3dfa2326af3443011c0de99ab6c47a22396809