rsyslog: have rsyslog create its own files
Prior to this change, puppet was responsible for creating the initial
log file for the systemd::syslog define type. After creation,
`replace => false` ensured the log content was not clobbered.
Unfortunately, puppet still calculates the hash of the file content
using Digest::Base, creating longer puppet runs. When possible the
use of `replace => false should be avoided.

After this change rsyslog creates the files itself. We also convert the
template to an epp, for better type checking and readability.

Hosts: basic
Change-Id: I32a06131094c6d01db74b7644a7e591722932674