[DPE Postgres] Only backup the latest postgres dump file
We have changed the way that the PostgreSQL dump works, so that the
most recent dump file is always called psql-all-dbs-latest.sql.gz

The reason for doing so is so that we can use bacula in a more
space-efficient way. This commit causes bacula only to retrieve the
latest file from the postgres server during the backup. Since the
filename doesn't change between backups, bacula can use a more efficient
algorithm to store differences between versions.

Bug: T316655
Change-Id: I51f72d380295978ae8e121f2d8f7cfe14d0c80c7