superset: fix database connection test for our mysql DBs
* Why
After the upgrade to superset 3.1.0, I started to notice that the test
database API call was failing for both our `wikishared` and
`mysql_staging` databases, even though the password stored in our
password_mapping dict worked and allowed me to connect to the DB via the
mariadb CLI.

It turns out that the test database command calls the password lookup
function twice:
- once to get the password in the fist place, that is immediately
  obfuscated, to avoid leaking it in the logs
- a second time to de-obfuscate the password, right before we actually
  connect to the database.

When the password lookup function is called the second time, the URI now
contains the obfuscated password, and looks like this
mysql://research:XXXXXXXXXX@dbstore1009.eqiad.wmnet:3320/wikishared

As this exact key does not appear in our password mapping, the password
lookup function returns None, and the connection fails.

* What
We strip the obfuscated password from the URI to make sure the same
password is found and returned for both calls.

Bug: T335356
Change-Id: I494f7517daa2d6f4835998bcb11b456e354be566