Add a new Icinga contact group for team-data-platform
We need to differentiate between alerts that should be sent to the
data-platform-sre team and alerts that should be sent to the data
engineers responsible for data pipelines.

As such, we are creating a new set of communication channels for the
data-platform-sre team to use. This new contactgroup in icinga is part
of that change.

I will add the two specific contacts into the puppet private repository
before merging this change.

Bug: T342578
Change-Id: I42c2f0343c22cbdcfd0a9fc23c55a112b883d682