Update contactgroups for analytics => team-data-platform icinga checks
We have already updated the default role contacts for all hosts that
previously notified the analytics contactgroup in Icinga, so that they
now update the team-data-platform contactgroup instead.

This change now updates those checks that are specifically defined
within the puppet manifests, so that the majority switch to the new
contactgroup as well.

A number of them have been deemed to be specifically related to the data
pipelines and these will continue to notify the analytics contactgroup
as well as team-data-platform. These include varnishkafka, eventstreams,
and eventlogging checks.

Bug: T342578
Change-Id: I5ef4f2655f10304e8c5d480dc7fa7de6f1fe0461