Add wmcs-empty-rbd-trash script
This is a helper script to recursively delete things
from the rbd trash which might otherwise be 'stuck' due to
having associated images and snapshots.

If an image is in the trash but has a volume based off of a snapshot
which is /not/ in the trash then this could be destructive! I'm not
sure if that case is possible though.

Bug: T356904
Change-Id: I3ee683693b2bfcd1aa20cadb03351428967a0cb5