Change all role contacts for Data Engineering -> Data Platform
We have created a new team in Alertmanager to represent the Data
Platform SRE team. This commit updates the role contact for all of the
servers whose contact is currently set to Data Engineering.

The underlying reason for this change is so that we can use different
communication channels for monitoring data pipelines and the platform
itself.

Hosts: auto
Bug: T346438
Bug: T352783
Bug: T342578
Change-Id: I5d9fea9f79a406b65bf708260bdac99bab5ab103