grafana: Prevent race condition by excluding 'wal' directory in Loki sync
A race condition was discovered during the synchronization of Loki data
via the quickdatacopy service, triggered when the Loki instance on the
active host is writing to a WAL file while the quickdatacopy daemon
attempts to copy this file mid-write. This concurrent operation leads to
the corruption of the WAL file, causing the Loki daemon to fail. To
address this issue, the synchronization process excludes the 'wal'
directory, preventing the corruption of WAL files.

Bug: T357026
Change-Id: I9c3555ead0e1c5f35aa79a07da0e7ace899348af