idp_test: Register superset and superset-next IDP services
We register an OIDC test servics for the superset-next service,
currently being deployed to the K8s DSE cluster.

This differs from the current bare metal deployment where we rely on Apache<-->LDAP
to get a UNIX username to identify the user in Superset.

Note; we use the superset-next-k8s.eqiad.wmnet temporary domain,
as superset-k8s.wikimedia.org is currently production-bearing.

Bug: T353794
Change-Id: Ieb4224e72bfba225a068a3132657c9ec8d4ad3e5