package_builder: add hook for building HAProxy 2.6 component
We want to build HAProxy 2.6 against OpenSSL 1.1.1w given the possible
performance regressions in the version HAProxy ships with, OpenSSL 3.x.
To be able to do this, we use a separate component for haproxy that will
allow it to co-exist with the thirdparty/haproxy (deployed in
production).

Hosts: C:package_builder
Bug: T352744
Change-Id: I8f0a84b71da1bca382459a4873ce15666cbbaa4b