puppetdb: Use the nginx certs
Prior to this change we used the certs in /etc/puppetdb/ssl, but we
create certs in /etc/nginx/ssl that are never used. This doesn't matter
in practice since the certs are identical, as they are "exposed" puppet
certs. However, it is still confusing. Change to use the /etc/nginx/ssl
certs. Also, fix the nginx config to require the certs.

Hosts: C:puppetmaster::puppetdb
Bug: T342784
Change-Id: If8ebf541c1648312558f012356488267ec0377ae