convert ncmonitor role to puppet7
While debugging for T357449 I noticed these hosts
don't seem to be converted to puppet 7 yet, unlike most.

Realized when the reimage cookbook asked me which puppet version I want.

It's likely because ncmonitor doesn't appear on T349619 as it didn't
exist when that ticket was created.

edit: I noticed the hosts file in hiera after uploading PS1. So it's
just moving this to role level for another ncmonitor hosts besides 1001.

Bug: T349619
Change-Id: I6c181e4ab09b522f1b12f563f8d3ac9194509a39