D:service::uwsgi Parse the Icinga disable flag to uwsgi:app.
The uwsgi::app doesn't need to be monitored by Icinga, the systemd
check in Prometheus will trigger an alert if the service is not
running. Parse the Icinga check variable to the define, as this
should already indicate that the service shouldn't be monitored by
Icinga. This is also disable the check for all the uwsgi service
in task T350694.

Bug: T350694
Change-Id: Ie8be507f2ce98425d7f5474975e61a6f3066668a