Add support for routed Ganeti in D-I early_command.sh
as "netcfg" doesn't support a /32 netmask, keep the intial /23 then
quickly set it back to /32.

Tested on sretest2005.

This allows D-I to fetch packages and other files needed for the VM
config even if the proxy or target is on the same range.

Of course "early_command.sh" needs not to be in the same range to be
fetched.

Bug: T300152
Change-Id: I88ee6a272bc6ed006f8da2164b3da71464c66c9e