puppetserver: don't fail setting up rsync_module
Prior to this change we fail if the host is not the CA nor does it have
the puppetserver::ca class defined in puppetdb. This creates a chicken
and egg problem when standing up a new puppetserver, since on the first
puppet run the puppetserver::ca class will not be defined in puppetdb.

It is not clear why the additional check is needed since we only include
puppetserver::rsync_module from profile::puppetserver, so switch to a
warning.

Bug: T356991
Hosts: C:profile::puppetserver
Change-Id: I482b81e3b401d4be07e334920642b513ea4b012e